requirements.txt
streamlit
speechrecognition
pyaudio
mysql-connector-python
sqlalchemy
pandas